+++ src/pages/Dashboard/DashboardV1.js
@@ -14,7 +14,7 @@ import YieldToken from "abis/YieldToken.json";
 import "./Dashboard.css";
 
 import { t, Trans } from "@lingui/macro";
-import ExternalLink from "components/Common/ExternalLink";
+import ExternalLink from "components/ExternalLink/ExternalLink";
 import metamaskImg from "img/metamask.png";
 import coingeckoImg from "img/coingecko.png";
 import bscscanImg from "img/bscscan.png";
